processor/requirements.txt
python-dotenv
boto3
pika